setup.py for Misage 0.5.x
One of the strengths of Python is that you can install python packages in many
ways and use them in you project.

Unfortunately that is not the case with Misango. In stead of using standard
python tools and infrastructure one has to exception and for Misango and the
deployment in a different way.

This change fixes that by making Misago fully working Python package.

The package also includes data files. Data files can be accessed using
setuptools:

    import pkg_resources as pres

    MISAGO_DIR = pres.resource_filename('misago', '')

    STATICFILES_DIRS = (
        os.path.join(MISAGO_DIR, 'attachments')
        os.path.join(MISAGO_DIR, 'media')
        os.path.join(MISAGO_DIR, 'static')
    )