Add application HealthCheck (#1263)
At the current state of the application, it's kind of pointless to have
any sort of healthcheck endpoint present. This would however become
really useful, once we are considering running Misago on a self-healing
infrastructure, such as Kubernetes.

K8s has a functionality monitoring the [readiness and liveliness][1]
probes, expecting everything to be functioning correctly. If an
unexpected status is exposed via URL, it will try to kill the
application causing a recreation with hopefully healthy state.

The proposed addition, will check the django runtime, including the
database connection which is being touched by a django middleware before
the view is rendered. I hope, this will grow over time, with fields such
as `cache` or `queue`.

[1]: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-probes/